refactor(core): Move `typeorm` operators from `WorkflowService` to `WorkflowRepository` (no-changelog)
